Switch from double to decimal for financial calculations
Replaced `double` with `decimal` across the codebase to improve precision and accuracy in financial calculations. Updated properties such as `FaceValue`, `YearsToMaturity`, `DiscountFactor`, `PresentValue`, and `Rate.Value`. Adjusted method signatures, internal logic, and test cases to reflect the change.

- Updated `CalculatePresentValue` methods to use `decimal` and added rounding logic for 4 decimal places.
- Modified `Rate.Parse` to handle `decimal` values and percentage strings.
- Replaced `FlexibleDoubleConverter` with `FlexibleDecimalConverter` for CSV parsing.
- Refactored test data and assertions to use `decimal` literals.
- Ensured backward compatibility by casting `decimal` to `double` where necessary (e.g., `Math.Pow`).

These changes align with best practices for financial applications, ensuring higher precision and reliability in bond valuation and related computations.

+++ Application/DTOs/BondValuationDTO.cs
@@ -4,10 +4,10 @@
     {
         public required string BondId { get; set; }
         public required string Type { get; set; }
-        public required double PresentValue { get; set; }
+        public required decimal PresentValue { get; set; }
         public required string Issuer { get; set; }
         public required string Rating { get; set; }
-        public required double YearsToMaturity { get; set; }
+        public required decimal YearsToMaturity { get; set; }
         public required string DeskNotes { get; set; }
     }
 }
