+++ DomainTests/BondsTests/BondTests.cs
@@ -12,12 +12,12 @@ namespace DomainTests.BondsTests
                 BondId = "BOND001",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.70%"), 
-                FaceValue = 1000.0,
+                FaceValue = 1000.0m,
                 PaymentFrequency = "annual",
                 Rating = "AAA",
                 Type = "Coupon",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 0.95
+                YearsToMaturity = 10.0m,
+                DiscountFactor = 0.95m
             };
         }
 
@@ -30,19 +30,19 @@ namespace DomainTests.BondsTests
             // Assert
             Assert.Equal("BOND001", bond.BondId);
             Assert.Equal("US Treasury", bond.Issuer);
-            Assert.Equal(0.057, bond.Rate.Value);
-            Assert.Equal(1000.0, bond.FaceValue);
+            Assert.Equal(0.057m, bond.Rate.Value);
+            Assert.Equal(1000.0m, bond.FaceValue);
             Assert.Equal("annual", bond.PaymentFrequency);
             Assert.Equal("AAA", bond.Rating);
             Assert.Equal("Coupon", bond.Type);
-            Assert.Equal(10.0, bond.YearsToMaturity);
-            Assert.Equal(0.95, bond.DiscountFactor);
+            Assert.Equal(10.0m, bond.YearsToMaturity);
+            Assert.Equal(0.95m, bond.DiscountFactor);
         }
     }
 
     public class TestBond : Bond
     {
-        public override double CalculatePresentValue()
+        public override decimal CalculatePresentValue()
         {
             // Simple implementation for testing base class functionality
             return FaceValue * DiscountFactor;
