+++ DomainTests/ZeroCouponTests.cs
@@ -9,12 +9,12 @@
                 BondId = "ZERO001",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.00%"), // 5% annual rate
-                FaceValue = 1000.0,
+                FaceValue = 1000m,
                 PaymentFrequency = "none", // Zero-coupon typically has no periodic payments
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 1.0 // No additional discount for simplicity
+                YearsToMaturity = 10m,
+                DiscountFactor = 1m // No additional discount for simplicity
             };
         }
 
@@ -27,7 +27,7 @@
             // Expected calculation: PV = (1 + 0.05)^10 × 1000 × 1.0
             // (1.05)^10 ≈ 1.628894626777441
             // 1000 × 1.628894626777441 ≈ 1628.89
-            double expectedPresentValue = 1628.89;
+            decimal expectedPresentValue = 1628.89m;
 
             // Act
             var result = bond.CalculatePresentValue();
@@ -45,16 +45,16 @@
                 BondId = "ZERO123",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.00%"), 
-                FaceValue = 1000.0,
+                FaceValue = 1000m,
                 PaymentFrequency = "none",
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 0.9 // 10% discount
+                YearsToMaturity = 10m,
+                DiscountFactor = 0.9m // 10% discount
             };
 
             // Expected: (1.05)^10 × 1000 × 0.9 ≈ 1628.89 × 0.9 = 1466.01
-            double expectedPresentValue = 1466.01;
+            decimal expectedPresentValue = 1466.01m;
 
             // Act
             var result = bond.CalculatePresentValue();
@@ -72,19 +72,19 @@
                 BondId = "ZERO999",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.00%"), 
-                FaceValue = 0.0,
+                FaceValue = 0m,
                 PaymentFrequency = "none",
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 1.0
+                YearsToMaturity = 10m,
+                DiscountFactor = 1m
             };
 
             // Act
             var result = bond.CalculatePresentValue();
 
             // Assert
-            Assert.Equal(0.0, result);
+            Assert.Equal(0m, result);
         }
 
         [Fact]
@@ -96,19 +96,19 @@
                 BondId = "ZERO000",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.00%"), 
-                FaceValue = 1000.0,
+                FaceValue = 1000m,
                 PaymentFrequency = "none",
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 0.0
+                YearsToMaturity = 10m,
+                DiscountFactor = 0m
             };
 
             // Act
             var result = bond.CalculatePresentValue();
 
             // Assert
-            Assert.Equal(0.0, result);
+            Assert.Equal(0m, result);
         }
 
         [Fact]
@@ -120,16 +120,16 @@
                 BondId = "ZERO001",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("5.00%"),
-                FaceValue = 1000.0,
+                FaceValue = 1000m,
                 PaymentFrequency = "none",
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 0.0,
-                DiscountFactor = 0.95
+                YearsToMaturity = 0m,
+                DiscountFactor = 0.95m
             };
 
             // Expected: (1.05)^0 × 1000 × 0.95 = 1 × 1000 × 0.95 = 950.00
-            double expectedPresentValue = 950.00;
+            decimal expectedPresentValue = 950.00m;
 
             // Act
             var result = bond.CalculatePresentValue();
@@ -147,16 +147,16 @@
                 BondId = "ZERO002",
                 Issuer = "US Treasury",
                 Rate = Rate.Parse("0.00%"), // 0% annual rate
-                FaceValue = 1000.0,
+                FaceValue = 1000m,
                 PaymentFrequency = "none",
                 Rating = "AAA",
                 Type = "Government",
-                YearsToMaturity = 10.0,
-                DiscountFactor = 0.9
+                YearsToMaturity = 10m,
+                DiscountFactor = 0.9m
             };
 
             // Expected: (1 + 0)^10 × 1000 × 0.9 = 1 × 1000 × 0.9 = 900.00
-            double expectedPresentValue = 900.00;
+            decimal expectedPresentValue = 900m;
 
             // Act
             var result = bond.CalculatePresentValue();
